+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.3] - 2025-07-27
+
 ### Added
 - Add trusted publishing
 - Add Zarr group functionality
@@ -210,7 +212,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial public release
 
-[unreleased]: https://github.com/LDeakin/zarrs_ffi/compare/v0.9.2...HEAD
+[unreleased]: https://github.com/LDeakin/zarrs_ffi/compare/v0.9.3...HEAD
+[0.9.3]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.9.3
 [0.9.2]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.9.2
 [0.9.1]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.9.1
 [0.9.0]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.9.0